Move In-Game Updates to Billing module.
