+++ android-appcoins-billing/src/main/java/com/appcoins/sdk/billing/usecases/ingameupdates/GetVanillaDeepLink.kt
@@ -1,4 +1,4 @@
-package com.appcoins.sdk.ingameupdates.usecases
+package com.appcoins.sdk.billing.usecases.ingameupdates
 
 object GetVanillaDeepLink {
