--- appcoins-in-game-updates/src/main/AndroidManifest.xml
@@ -1,4 +1,0 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.appcoins.sdk.ingameupdates" >
-  <uses-permission android:name="android.permission.INTERNET"/>
-</manifest>
