--- appcoins-in-game-updates/src/main/java/com/appcoins/sdk/ingameupdates/api/InGameUpdatesClient.kt
@@ -1,6 +1,0 @@
-package com.appcoins.sdk.ingameupdates.api
-
-interface InGameUpdatesClient {
-    fun isAppUpdateAvailable(): Boolean
-    fun launchAppUpdateFlow()
-}
