--- appcoins-in-game-updates/src/main/java/com/appcoins/sdk/ingameupdates/services/ServiceResponseListener.kt
@@ -1,5 +1,0 @@
-package com.appcoins.sdk.ingameupdates.services
-
-interface ServiceResponseListener {
-    fun onResponseReceived(requestResponse: RequestResponse?)
-}
