--- appcoins-in-game-updates/src/main/java/com/appcoins/sdk/ingameupdates/utils/ServiceUtils.kt
@@ -1,6 +1,0 @@
-package com.appcoins.sdk.ingameupdates.utils
-
-object ServiceUtils {
-    fun isSuccess(code: Int): Boolean =
-        code in 200..299
-}
