+++ settings.gradle
@@ -10,4 +10,3 @@ include ':appcoins-ads'
 include ':toolbox'
 include ':communication'
 include ':appcoins-lifecycle'
-include ':appcoins-in-game-updates'
